backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
mysqlclient==2.2.4
bcrypt==4.1.2
python-dotenv==1.0.1
gunicorn==21.2.0
